Small correction proposed as a path to video 45.
UPDATE: Fix for testing password too short
There is an issue with the previous lesson in the test: test_password_too_short  in app/user/tests/test_user_api.py

The issue is that the test will pass even if the password is longer than 5 characters.

This is because the name is required, and we don't provide it in the payload.

Therefore, even if a password of the correct length is provided, the API will still return HTTP 400 due to the missing name.

To correct this, update the payload to:

         payload = {
            'email': 'test@londonappdev.com',
            'password': 'pw',
            'name': 'Test',
        }
See the full change here.

The same issue is in test_user_exists, which can be solved by making the payload the same as above (see full change here).